requirements.txt
openai~=0.27.8
requests~=2.31.0

Pillow~=9.5.0